+++ syscall_cgo_linux.go
@@ -5,7 +5,9 @@
 
 package purego
 
-import "github.com/ebitengine/purego/internal/unknown"
+import (
+	"github.com/ebitengine/purego/internal/cgo"
+)
 
 // this is linked to a C function from internal/unknown
 var syscall9XABI0 uintptr
@@ -19,7 +21,7 @@ type syscall9Args struct {
 
 //go:nosplit
 func syscall_syscall9X(fn, a1, a2, a3, a4, a5, a6, a7, a8, a9 uintptr) (r1, r2, err uintptr) {
-	return unknown.Syscall9X(fn, a1, a2, a3, a4, a5, a6, a7, a8, a9)
+	return cgo.Syscall9X(fn, a1, a2, a3, a4, a5, a6, a7, a8, a9)
 }
 
 func NewCallback(_ interface{}) uintptr {
